move2Acc example datasets

move2Acc provides two example datasets for working with acceleration data. These data are publicly available and downloaded from Movebank.

See the Movebank Attribute Dictionary for details on data attributes.

Galapagos albatrosses

GPS and acceleration data for Galapagos albatrosses from Movebank study 2911040. Waved albatrosses were tracked during breeding and non-breeding periods between 2008 and 2010. Acceleration data are provided in burst format from e-obs tags.

Format: A move2 with 9 tracks and 54 features lasting from 2008-07-27 00:00:00 UTC to 2008-07-27 01:00:00 UTC.

Source: https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study2911040

Lesser black-backed gulls

GPS and acceleration data for lesser black-backed gulls (Larus fuscus, Laridae) breeding at the southern North Sea coast in Belgium and the Netherlands. Published by the Research Institute for Nature and Forest (INBO). Data collected by the LifeWatch GPS tracking network for large birds for the project/study LBBG_ZEEBRUGGE. Acceleration data are provided in long format from trackers developed by the University of Amsterdam Bird Tracking System (UvA-BiTS).

Format: A move2 with 5 tracks and 1499 features lasting from 2021-03-03 00:00:00 UTC to 2021-03-04 00:00:00 UTC.

Source: https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study985143423

Usage

albatrosses()

gulls()

Value

move2 object with GPS and acceleration data columns