View standard Movebank IMU data column sets

Movebank has several standard ways to store data for each IMU sensor. These functions show the recognized columns for each sensor that can be extracted from a move2 object by default.

movebank_acc_colsets() — standard column sets for as_acc().

movebank_mag_colsets() — standard column sets for as_mag().

movebank_gyro_colsets() — standard column sets for as_gyro().

To extract IMU data from a move2 with column names that don't correspond to Movebank's conventions, provide a custom set of IMU columns with imu_colset().

Usage

movebank_acc_colsets()

movebank_mag_colsets()

movebank_gyro_colsets()

Value

A named list of imu_colset objects.

Details

move2 objects store IMU data in two ways:

Long-format columns store one measurement (possibly for multiple axes) in a single row.

Burst-format columns store a burst of measurements as a space-delimited string. This string must be segmented into axis-specific measurements using an associated column that indicates the axes present for the bursted data. A further column provides the sampling frequency of the burst. All three of these columns must be present to form a valid burst-format column set.

See also

Examples

movebank_acc_colsets()
#> $eobs
#> burst-format [axes=eobs_acceleration_axes, frequency=eobs_acceleration_sampling_frequency_per_axis, bursts=eobs_accelerations_raw]
#> 
#> $burst
#> burst-format [axes=acceleration_axes, frequency=acceleration_sampling_frequency_per_axis, bursts=accelerations_raw]
#> 
#> $xyz
#> long-format [X=acceleration_x, Y=acceleration_y, Z=acceleration_z]
#> 
#> $raw_xyz
#> long-format [X=acceleration_raw_x, Y=acceleration_raw_y, Z=acceleration_raw_z]
#> 
movebank_mag_colsets()
#> $burst
#> burst-format [bursts=magnetic_fields_raw, axes=magnetic_field_axes, frequency=magnetic_field_sampling_frequency_per_axis]
#> 
#> $xyz
#> long-format [X=magnetic_field_x, Y=magnetic_field_y, Z=magnetic_field_z]
#> 
#> $raw_xyz
#> long-format [X=magnetic_field_raw_x, Y=magnetic_field_raw_y, Z=magnetic_field_raw_z]
#> 
movebank_gyro_colsets()
#> $burst
#> burst-format [bursts=angular_velocities_raw, axes=gyroscope_axes, frequency=gyroscope_sampling_frequency_per_axis]
#> 
#> $xyz
#> long-format [X=angular_velocity_x, Y=angular_velocity_y, Z=angular_velocity_z]
#>