Plot bursts over time

Plot the trace of IMU values from an IMU vector with time on the x-axis.

Usage

plot_time(x, ylab = "Value")

Arguments

x

An IMU vector (e.g. acc, mag, gyro).

ylab

A character with the y axis label

Details

If the bursts in the input come from multiple sources, traces may be combined incorrectly. See examples.

Examples

plot_time(acc_example())

# If bursts come from multiple sources (in this case, deployments),
# then lines from different bursts may be incorrectly connected:
alb <- albatrosses()
a <- as_acc(alb)

plot_time(a)

# To avoid this issue, plot only a single deployment's values:
plot_time(a[move2::mt_track_id(alb) == "4261-2228"])